chore: remove Flags.json
