+++ src/flags.ts
@@ -145,20 +145,6 @@ export const help = (opts: Partial<BooleanFlag<boolean>> = {}): BooleanFlag<void
   },
 })
 
-/**
- * Boolean flag for enabling JSON output. We strongly discourage using this flag directly
- * and instead recommend setting the `enableJsonFlag` property on the Command class.
- */
-export const json = (opts: Partial<BooleanFlag<boolean>> = {}): BooleanFlag<boolean> => boolean({
-  description: 'Format output as json.',
-  helpGroup: 'GLOBAL',
-  ...opts,
-  async parse(input, {config}) {
-    if (input) process.env[config.scopedEnvVarKey('JSON_FLAG_OVERRIDE')] = 'true'
-    return input
-  },
-})
-
 type ElementType<T extends ReadonlyArray<unknown>> = T[number];
 
 export function option<T extends readonly string[], P extends CustomOptions>(
